Add files via upload
Introduction to the Rock Paper Scissors Game
Before building any new game, one needs to understand all about the game including its rules, inputs, and outputs, and all the necessary things.

How to play the game?
Rock paper scissors is one of the classic hand games we have been playing since we were all small kids. It is typically played between two players (can be played with more than two players as well) each one forming one of three shapes (rock, paper, and scissors) at the same time, with an extended hand.
Rules of the Game
Although the coding part of the game can get a little complex the game rules are simple which are as follows:

Rock wins against scissors.
Scissors win against the paper.
Paper wins against the rock.
Building Rock Paper Scissors in Python
We will divide the whole game building into parts to make it simple for you! All the steps we will be taking for building the game are mentioned below:

Importing all the necessary modules
Printing Greeting Messages and asking the player about how it wants to play the game.
Creating separate functions for the following purposes:
Draw the Hand Diagrams
Play with the Computer!
Play with a Friend!